app/authentication/urls.py
from django.urls import path
from authentication import views
from rest_framework.authtoken.views import obtain_auth_token

urlpatterns = [
    path('register/', views.RegisterView.as_view(), name='register'),
    path('login/', views.UserLogin.as_view(), name='login'),
]