app/requirements.txt
Django==4.2.3
djangorestframework